PdfEncodingFactory: Try fix compilation with gcc

+++ src/podofo/main/PdfEncodingFactory.cpp
@@ -168,14 +168,14 @@ PdfEncoding PdfEncodingFactory::CreateEncoding(const PdfDictionary& fontDict, co
     if (encoding->GetType() != PdfEncodingMapType::CMap)
     {
         // Try read limits /FirstChar and /LastChar for simple encodings
-        // NOTE: Simple fonts are ;imited to one-byte encodings, hence
+        // NOTE: Simple fonts are limited to one-byte encodings, hence
         // read values can be clamped to [0, 255]
         int64_t num;
         if (fontDict.TryFindKeyAs("FirstChar", num))
-            parsedLimits.FirstChar = PdfCharCode((unsigned)std::clamp(num, 0LL, 255LL), 1);
+            parsedLimits.FirstChar = PdfCharCode((unsigned)std::clamp((int)num, 0, 255), 1);
 
         if (fontDict.TryFindKeyAs("LastChar", num))
-            parsedLimits.LastChar = PdfCharCode((unsigned)std::clamp(num, 0LL, 255LL), 1);
+            parsedLimits.LastChar = PdfCharCode((unsigned)std::clamp((int)num, 0, 255), 1);
 
         if (parsedLimits.LastChar.Code >= parsedLimits.FirstChar.Code)
         {
